JANUS-257: Fixed weird tile_id values

+++ src/ska_low_mccs_daq/pydaq/daq_receiver_interface.py
@@ -62,9 +62,9 @@ class DaqReceiver:
 
     class ChannelMetadata(ctypes.Structure):
         _fields_ = [
-            ("tile_id", ctypes.c_int16),
-            ("cont_channel_id", ctypes.c_int32),
-            ("nof_packets", ctypes.c_uint32),
+            ("tile_id", ctypes.c_uint8),
+            ("cont_channel_id", ctypes.c_uint32),
+            ("nof_packets", ctypes.c_uint64),
             ("packet_counter", ctypes.c_uint32 * 2048),
             ("payload_length", ctypes.c_uint64),
             ("sync_time", ctypes.c_uint64),
